+++ tmp-agg/Mini.Tests.ps1
@@ -1,0 +1,1 @@
+Describe \ Mini\ -Tag Slow { It \x\ { 1 | Should -Be 1 } }